+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * [HexaPDF::Task::MergeAcroForm] for merging AcroForm information for imported
   pages
-* [HexaPDF::Document#write_to_string] for easily writing a document to a String
+* [HexaPDF::Document#write_to_string] and [HexaPDF::Composer#write_to_string]
+  for easily writing a document to a String
 
 ### Changed